Added MoonWarriors and Watermelon resources for Mac

+++ tests/SpiderMonkeyTest/ScriptingCore.h
@@ -122,7 +122,7 @@ extern "C" {
 	void set_proxy_for_jsobject(JSPROXY_NSObject* proxy, JSObject *jsobj);
 	void del_proxy_for_jsobject(JSObject *jsobj);
 	
-	JSBool set_reserved_slot(JSObject *obj, NSUInteger idx, jsval value);
+	JSBool set_reserved_slot(JSObject *obj, uint32_t idx, jsval value);
 	
 #ifdef __cplusplus
 }
